Maps extracted to its own Folder Styles extract to JSON
