+++ frontend/src/Editor/Box.jsx
@@ -14,7 +14,7 @@ import { DaterangePicker } from './Components/DaterangePicker';
 import { Multiselect } from './Components/Multiselect';
 import { Modal } from './Components/Modal';
 import { Chart } from './Components/Chart';
-import { Map } from './Components/Map';
+import { Map } from './Components/Map/Map';
 import { renderTooltip } from '../_helpers/appUtils';
 import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
